# openapi.yaml
openapi: 3.0.3
info:
  title: QA Retail — API Backend
  description: >
    Contrato de la API REST de "ChileRetail", la tienda demo usada como campo de
    practica de QA. Documenta, en formato contract-first (OpenAPI), el
    comportamiento real de cada endpoint ya implementado en `src/routes/*.routes.js`:
    parametros, cuerpos de peticion, respuestas y codigos de estado (incluyendo los
    casos negativos sembrados a proposito para practicas de testing).

    Reglas transversales:
    - Todas las rutas estan bajo el prefijo `/api` (ver `servers`).
    - Las rutas marcadas con el candado requieren `Authorization: Bearer <token>`
      (JWT, expira en 2h). Sin token o token invalido/expirado -> **401**.
    - Las rutas marcadas ademas como "admin" requieren `role: admin` en el token.
      Autenticado pero sin rol admin -> **403**.
    - Toda ruta no definida en este contrato responde **404**
      `{ "error": "Ruta no encontrada" }`.
    - Cualquier excepcion no controlada responde **500**
      `{ "error": "Error interno del servidor" }`.
  version: "1.0.0"
  contact:
    name: QA Retail — Backend
servers:
  - url: http://localhost:4000/api
    description: Servidor local de desarrollo
...
security: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Registra un nuevo usuario (rol customer)
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: >
            Faltan campos obligatorios, email con formato invalido, o password
            con menos de 6 caracteres.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: El email ya esta registrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/login:
    post:
      tags: [auth]
      summary: Inicia sesion y obtiene un JWT
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Faltan email o password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Email no registrado o password incorrecto.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/me:
    get:
      tags: [auth]
      summary: Perfil del usuario autenticado
      operationId: getMe
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Perfil del usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: El usuario del token ya no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [auth]
      summary: Actualiza nombre, telefono, RUT y/o fecha de nacimiento
      description: >
        El email, el rol y los puntos de lealtad **no** son editables por esta via.
        Cada campo es opcional e independiente; enviar `null` o `""` en `phone`,
        `rut` o `birthDate` los limpia.
      operationId: updateMe
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Perfil actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: >
            `name` vacio, `phone` que no cumple `+56 9 XXXX XXXX`, `rut` invalido
            (modulo 11) o `birthDate` con formato distinto de `YYYY-MM-DD`,
            futura, o de mas de 120 anios de antiguedad.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: El usuario del token ya no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/password:
    put:
      tags: [auth]
      summary: Cambia la contraseña del usuario autenticado
      operationId: updatePassword
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "200":
          description: Contraseña actualizada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                required: [ok]
        "400":
          description: >
            Faltan `currentPassword`/`newPassword`, o `newPassword` tiene menos
            de 6 caracteres.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Sin token, token invalido, o `currentPassword` incorrecta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: El usuario del token ya no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"

  # -------------------------------------------------------------- products --
...
  /orders:
    get:
      tags: [orders]
      summary: Lista las ordenes del usuario autenticado
      operationId: listMyOrders
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Ordenes del usuario (vacia si no tiene ninguna).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    get:
      tags: [orders]
      summary: Obtiene una orden por id
      description: >
        **Nota de contrato:** la implementacion actual no filtra por
        propietario — cualquier usuario autenticado que conozca el `id` puede
        leer la orden. Documentado para que se evalue/pruebe explicitamente
        (candidato a control de autorizacion adicional).
      operationId: getOrderById
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Orden encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # --------------------------------------------------------------- reviews --
...
  /addresses:
    get:
      tags: [addresses]
      summary: Lista las direcciones del usuario autenticado
      operationId: listAddresses
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Direcciones del usuario (vacia si no tiene ninguna).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [addresses]
      summary: Crea una direccion
      operationId: createAddress
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressCreateRequest"
      responses:
        "201":
          description: Direccion creada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          description: Faltan `label`, `street`, `city` o `region`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /addresses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    put:
      tags: [addresses]
      summary: Actualiza una direccion propia
      description: Todos los campos son opcionales; solo se aplican los enviados.
      operationId: updateAddress
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressUpdateRequest"
      responses:
        "200":
          description: Direccion actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: La direccion no existe o no pertenece al usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [addresses]
      summary: Elimina una direccion propia
      operationId: deleteAddress
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Direccion eliminada.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: La direccion no existe o no pertenece al usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"

  # ----------------------------------------------------------- admin-orders --
  /admin/orders:
    get:
      tags: [admin-orders]
      summary: Lista todas las ordenes de todos los usuarios (admin)
      operationId: adminListOrders
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Todas las ordenes del sistema.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /admin/orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    get:
      tags: [admin-orders]
      summary: Obtiene cualquier orden por id (admin)
      operationId: adminGetOrderById
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Orden encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /admin/orders/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    put:
      tags: [admin-orders]
      summary: Actualiza el estado de una orden (admin)
      description: >
        Flujo de negocio esperado: `pagado -> enviado -> entregado` (o
        `cancelado` desde `pagado`/`enviado`). **Nota de contrato:** la
        implementacion actual no valida el valor de `status` contra ese flujo
        ni contra un enum — acepta cualquier string no vacio. Buen candidato
        para pruebas de validacion de entrada.
      operationId: adminUpdateOrderStatus
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdateRequest"
      responses:
        "200":
          description: Orden actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # --------------------------------------------------------------- shipping --
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT devuelto por `POST /auth/register` o `POST /auth/login`. Payload:
        `{ id, email, role }`. Expira en 2 horas.
  responses:
    Unauthorized:
      description: Sin token (`Token no proporcionado`) o token invalido/expirado (`Token invalido o expirado`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Autenticado, pero sin rol admin (`Requiere rol admin`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Error interno no controlado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example: { error: "Mensaje descriptivo del error" }
...
    Role:
      type: string
      enum: [customer, admin]
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string, format: email, example: cliente@demo.cl }
        name: { type: string, example: Cliente Demo }
        phone:
          type: string
          nullable: true
          example: "+56 9 1234 5678"
        rut:
          type: string
          nullable: true
          example: "18.420.789-3"
        birthDate:
          type: string
          format: date
          nullable: true
          example: "1992-10-14"
        role:
          $ref: "#/components/schemas/Role"
        loyaltyPoints: { type: integer, example: 1450 }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, phone, rut, birthDate, role, loyaltyPoints, createdAt]
    RegisterRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        name: { type: string }
      required: [email, password, name]
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: JWT firmado, expira en 2h. }
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    UpdateMeRequest:
      type: object
      description: Todos los campos son opcionales.
      properties:
        name: { type: string }
        phone:
          type: string
          nullable: true
          description: 'Formato `+56 9 XXXX XXXX` (acepta variantes sin "+" o espacios). `null`/`""` limpia el campo.'
          example: "+56 9 8765 4321"
        rut:
          type: string
          nullable: true
          description: Validado con algoritmo modulo 11. `null`/`""` limpia el campo.
          example: "18420789K"
        birthDate:
          type: string
          nullable: true
          description: 'Formato `YYYY-MM-DD`, no futura, no mayor a 120 años. `null`/`""` limpia el campo.'
          example: "1992-10-14"
    UpdatePasswordRequest:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 6 }
      required: [currentPassword, newPassword]
...
    OrderItem:
      type: object
      description: Snapshot del producto al momento de la compra (no referencia al catalogo vivo).
      properties:
        productId: { type: integer }
        name: { type: string }
        unitPrice: { type: number }
        quantity: { type: integer }
      required: [productId, name, unitPrice, quantity]
    Order:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal: { type: number }
        discount: { type: number }
        couponApplied:
          type: string
          nullable: true
        tax: { type: number }
        shipping: { type: number }
        total: { type: number }
        status:
          type: string
          description: 'Valor inicial siempre "pagado"; actualizable via `PUT /admin/orders/{id}/status`.'
          example: pagado
        createdAt: { type: string, format: date-time }
      required:
        - id
        - userId
        - items
        - subtotal
        - discount
        - couponApplied
        - tax
        - shipping
        - total
        - status
        - createdAt
    OrderStatusUpdateRequest:
      type: object
      properties:
        status:
          type: string
          description: Sin enum impuesto por el servidor; valores de negocio esperados a continuacion.
          enum: [pagado, enviado, entregado, cancelado]
      required: [status]
...
    Address:
      type: object
      properties:
        id: { type: integer }
        label: { type: string, example: Casa }
        street: { type: string }
        city: { type: string }
        region: { type: string }
        isDefault: { type: boolean }
      required: [id, label, street, city, region, isDefault]
    AddressCreateRequest:
      type: object
      properties:
        label: { type: string }
        street: { type: string }
        city: { type: string }
        region: { type: string }
        isDefault: { type: boolean, default: false }
      required: [label, street, city, region]
    AddressUpdateRequest:
      type: object
      description: Todos los campos son opcionales; solo se aplican los enviados.
      properties:
        label: { type: string }
        street: { type: string }
        city: { type: string }
        region: { type: string }
        isDefault: { type: boolean }
...